+++ Movie.java
@@ -1,7 +1,7 @@
 public class Movie extends AMedia {
 
 
-    public Movie(String name, int releaseYear, String category, int rating) {
+    public Movie(String name, int releaseYear, String category, float rating) {
         super(name, releaseYear, category, rating);
     }
